// src/app/api/api/models/cliente-response.ts
/* tslint:disable */
/* eslint-disable */
export interface ClienteResponse {
  activo?: boolean;
  clienteId?: number;
  contacto?: string;
  correo?: string;
  direccion?: string;
  estadoDsc?: string;
  nombreComercial?: string;
  numeroDocumento?: string;
  pais?: string;
  razonSocial?: string;
  telefono?: string;
  tipoDocumento?: string;
}