// src/app/api/api/models/cliente-registro-request.ts
/* tslint:disable */
/* eslint-disable */
export interface ClienteRegistroRequest {
  clienteId: number;
  contacto?: string;
  correo?: string;
  direccion?: string;
  nombreComercial?: string;
  numeroDocumento: string;
  pais: string;
  razonSocial: string;
  telefono?: string;
  tipoDocumento: string;
}